openroad_analytics/requirements.txt
Cerberus==1.3.4
certifi==2021.10.8
click==8.0.3
dnspython==1.16.0
Eve==1.1.5
Eve-Swagger==0.1.4
eventlet==0.30.2
Events==0.3
Flask==2.0.2
Flask-Classful==0.14.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
graphviz==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
haversine==2.5.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mongoengine==0.23.1
numpy==1.21.4
pandas==1.3.4
passlib==1.7.4
PyJWT==2.3.0
pymongo==3.12.1
pyproj==3.3.0
python-dateutil==2.8.2
python-statemachine==0.8.0
pytz==2021.3
Shapely==1.8.0
simplejson==3.17.6
six==1.16.0
undecorated==0.3.0
Werkzeug==2.0.2
